refactor: move nvcc-incompatible variant to separate mesh_utils file
fix: namespace for mesh functions
