+++ src/serac/numerics/CMakeLists.txt
@@ -9,6 +9,7 @@ set(numerics_headers
     expr_template_impl.hpp
     expr_template_ops.hpp
     mesh_utils.hpp
+    mesh_utils_base.hpp
     vector_expression.hpp
     assembled_sparse_matrix.hpp
     )
